+++ src/types.ts
@@ -1,4 +1,9 @@
-import { ApprovalTypes, FormTypes, SubmissionTypes } from '@oneblink/types'
+import {
+  ApprovalTypes,
+  EmailTemplateTypes,
+  FormTypes,
+  SubmissionTypes,
+} from '@oneblink/types'
 import jwksClient from 'jwks-rsa'
 export * from '@oneblink/types'
 
@@ -82,4 +87,14 @@ export type FormApprovalFlowInstanceResponse = {
   formApprovalFlowInstance: ApprovalTypes.FormApprovalFlowInstance
   form: FormTypes.Form
   formSubmissionApprovals: ApprovalTypes.FormSubmissionApproval[]
+}
+
+export type EmailTemplatesSearchResult = {
+  formsAppEnvironments: EmailTemplateTypes.EmailTemplate[]
+} & BaseSearchResult
+
+export type EmailTemplatesSearchOptions = {
+  formsAppEnvironmentId: number
+  limit?: number
+  offset?: number
 }